requirements.txt
Flask==2.0.1
Pillow==9.5.0
numpy==1.24.3
Werkzeug==2.0.2
Jinja2==3.0.3
MarkupSafe==2.0.1
itsdangerous==2.0.1
click==8.0.4 